requirements.txt
python-dotenv==1.0.1
discord.py==2.3.2
web3==6.15.1
redis==5.0.3
fastapi==0.110.0
uvicorn==0.28.0
python-json-logger==2.0.7